+++ SIGpi_installer.sh
@@ -175,6 +175,7 @@ select_gnuradio() {
         "Select a GNUradio version" 20 80 12 \
 		"gnuradio38" "GNU Radio 3.8 " ON \
         "gnuradio39" "GNU Radio 3.9 " OFF \
+        "gnuradio310" "GNU Radio 3.10 " OFF \
         3>&1 1>&2 2>&3)
     RET=$?
     if [ $RET -eq 1 ]; then
@@ -285,8 +286,7 @@ if [ "$1" = "node" ]; then
     source $SIGPI_PACKAGES/pkg_dump1090 install
     source $SIGPI_PACKAGES/pkg_multimon-ng install
     source $SIGPI_PACKAGES/pkg_radiosonde install
-
-    source $SIGPI_PACKAGES/pkg_sigpi-node install
+    source $SIGPI_PACKAGES/pkg_sigpi-node install $2
 
     echo -e "${SIGPI_BANNER_COLOR}"
     echo -e "${SIGPI_BANNER_COLOR} ##"
@@ -302,103 +302,6 @@ if [ "$1" = "node" ]; then
     sudo reboot
     exit 0
 
-fi
-
-# Edge option invoked ?
-if [ "$1" = "edge" ]; then
-    calc_wt_size
-    
-    TERM=ansi whiptail --title "SigPi Edge Install" --clear --msgbox "Ready to Install" 12 120
-
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   System Update & Upgrade"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_RESET}"
-
-    sudo apt-get -y update
-    sudo apt-get -y upgrade
-
-    touch $SIGPI_CONFIG
-    echo "sigpi_edge" >> $SIGPI_CONFIG
-    cd $SIGPI_SOURCE
-
-    source $SIGPI_SCRIPTS/install_server_dependencies.sh
-    source $SIGPI_SCRIPTS/install_core_devices.sh
-
-    # Install Libraries
-    source $SIGPI_PACKAGES/pkg_libsigmf install
-    source $SIGPI_PACKAGES/pkg_liquid-dsp install
-    source $SIGPI_PACKAGES/pkg_libbtbb install
-    source $SIGPI_PACKAGES/pkg_libdab install
-    source $SIGPI_PACKAGES/pkg_sgp4 install
-    source $SIGPI_PACKAGES/pkg_aptdec install
-    
-    source $SIGPI_PACKAGES/pkg_nrsc5 install
-    source $SIGPI_PACKAGES/pkg_cm256cc install
-    source $SIGPI_PACKAGES/pkg_mbelib install
-    source $SIGPI_PACKAGES/pkg_serialdv install
-    source $SIGPI_PACKAGES/pkg_dsdcc install
-    source $SIGPI_PACKAGES/pkg_codec2 install
-
-    source $SIGPI_PACKAGES/pkg_sdrangel install
-    source $SIGPI_PACKAGES/pkg_fftwf-wisdom install
-    source $SIGPI_PACKAGES/pkg_sigpi-edge install
-
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   SIGpi Edge Installation Complete !!"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   System needs to reboot for all changes to occur"
-    echo -e "${SIGPI_BANNER_COLOR} ##   Reboot will begin in 15 seconsds unless CTRL-C hit"
-    echo -e "${SIGPI_BANNER_RESET}"
-    sleep 17
-    sudo sync
-    sudo reboot
-    exit 0
-fi
-
-# Base option invoked ?
-if [ "$1" = "base" ]; then 
-
-    TERM=ansi whiptail --title "SigPi Base Install" --clear --msgbox "Ready to Install" 12 120
-
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   System Update & Upgrade"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_RESET}"
-
-    sudo apt-get -y update
-    sudo apt-get -y upgrade
-
-    touch $SIGPI_CONFIG
-    echo "sigpi_desktop" >> $SIGPI_CONFIG    
-    cd $SIGPI_SOURCE
-
-    source $SIGPI_SCRIPTS/install_core_dependencies.sh
-    source $SIGPI_SCRIPTS/install_core_devices.sh
-    source $SIGPI_SCRIPTS/install_desktop-prep.sh
-    source $SIGPI_PACKAGES/pkg_rtl_433 install
-    source $SIGPI_PACKAGES/pkg_dump1090 install
-    source $SIGPI_PACKAGES/pkg_gqrx install
-    source $SIGPI_SCRIPTS/install_desktop-post.sh
-
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   Base Installation Complete !!"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR}"
-    echo -e "${SIGPI_BANNER_COLOR} ##"
-    echo -e "${SIGPI_BANNER_COLOR} ##   System needs to reboot for all changes to occur"
-    echo -e "${SIGPI_BANNER_COLOR} ##   Reboot will begin in 15 seconsds unless CTRL-C hit"
-    echo -e "${SIGPI_BANNER_RESET}"
-    sleep 17
-    sudo sync
-    sudo reboot
-    exit 0
 fi
 
 # Otherwise we are Full install
@@ -474,6 +377,10 @@ fi
 
 if grep gnuradio39 "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_gnuradio39 install
+fi
+
+if grep gnuradio310 "$SIGPI_INSTALLER"; then
+    source $SIGPI_PACKAGES/pkg_gnuradio310 install
 fi
 
 # gqrx
@@ -558,10 +465,8 @@ if grep audacity "$SIGPI_INSTALLER"; then
     source $SIGPI_PACKAGES/pkg_audacity install 
 fi
 
-# PAVU
-if grep pavu "$SIGPI_INSTALLER"; then
-    source $SIGPI_PACKAGES/pkg_pavucontrol install
-fi
+# PAVU  - Made Mandatory install in SIGpi 5.2 to support virtual audio cables
+source $SIGPI_PACKAGES/pkg_pavucontrol install
 
 # splat
 if grep splat "$SIGPI_INSTALLER"; then